depth_estimation/pyproject.toml
[tool.poetry]
name = "depth-estimation"
version = "0.1.0"
description = "Depth estimation project for the ENNCA course at Warsaw University of Technology, summer semester 23/24"
authors = ["Rubén J. <rubenjr@protonmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
data = "src.data:main"
net = "src.net:main"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.2"
torch = { version = "^2.0.0", source="torch"}
torchvision = { version = "^0.15.1", source="torch"}
pillow-simd = "^9.0.0.post1"

[[tool.poetry.source]]
name = "torch"
url = "http://download.pytorch.org/whl/cpu/"
secondary = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.259"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
